refactor: callbacks
Updates app to load callbacks on page load

Callbacks moved from halberd.py to dedicated pages

Restructuring prevents all callback to load on app initialization

Improves app performance

Restructuring also allows for better code management
