+++ pages/dashboard/recon_users.py
@@ -1,5 +1,6 @@
-import dash_bootstrap_components as dbc
 from dash import dcc,html
+import dash_bootstrap_components as dbc
+
 from core.entra.graph_request import GraphRequest
 
 # find role ID
